refactor unit tests for stability
